Make aac_cb_* tables uint8_t

+++ aacenc/aaccoder.c
@@ -71,8 +71,8 @@ static av_always_inline int quant2(float coef, const float Q)
     return av_clip((int)(pow(fabsf(coef) * Q, 0.75)), 0, 8191);
 }
 
-static const float aac_cb_range[12] = { 0, 3, 3, 3, 3, 9, 9, 8, 8, 13, 13, 17};
-static const float aac_cb_maxval[12] = {0, 1, 1, 2, 2, 4, 4, 7, 7, 12, 12, 16};
+static const uint8_t aac_cb_range [12] = {0, 3, 3, 3, 3, 9, 9, 8, 8, 13, 13, 17};
+static const uint8_t aac_cb_maxval[12] = {0, 1, 1, 2, 2, 4, 4, 7, 7, 12, 12, 16};
 
 /**
  * Calculate rate distortion cost for quantizing with given codebook
